requirements.txt
# Pirate Launcher — runtime dependencies
Pillow>=9.0
requests>=2.28.0
psutil>=5.9.0
aiohttp>=3.9.0
pyperclip>=1.8.0
tqdm>=4.65.0

# Optional: system tray (Settings → Preferences → Close to system tray)
pystray>=0.19.0
